Rename packages and update version of Jooby
