+++ pom.xml
@@ -1,111 +1,113 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
-  <modelVersion>4.0.0</modelVersion>
+    <modelVersion>4.0.0</modelVersion>
 
-  <parent>
-    <groupId>org.jooby</groupId>
-    <artifactId>jooby-project</artifactId>
-    <version>1.1.2</version>
-  </parent>
+    <parent>
+        <groupId>org.jooby</groupId>
+        <artifactId>jooby-project</artifactId>
+        <version>1.5.0</version>
+    </parent>
 
-  <artifactId>polymer-101-backend</artifactId>
-  <groupId>info.lefoll.jeff.polymer.workshop</groupId>
-  <version>1.0</version>
+    <artifactId>blog-formation-backend</artifactId>
+    <groupId>fr.avalonlab.blog.workshop</groupId>
+    <version>1.0</version>
 
-  <name>polymer-101-backend</name>
-  <description>generated by Jooby archetype</description>
+    <name>blog-formation-backend</name>
+    <description>generated by Jooby archetype</description>
 
-  <properties>
-    <jooby.version>1.1.2</jooby.version>
+    <properties>
+        <jooby.version>1.5.0</jooby.version>
+        <jackson.version>2.9.6</jackson.version>
 
-    <!-- Startup class -->
-    <application.class>info.lefoll.jeff.polymer.workshop.App</application.class>
-  </properties>
+        <!-- Startup class -->
+        <application.class>fr.avalonlab.blog.workshop.App</application.class>
+    </properties>
 
-  <dependencies>
-    <dependency>
-      <groupId>org.projectlombok</groupId>
-      <artifactId>lombok</artifactId>
-      <version>1.16.16</version>
-      <scope>provided</scope>
-    </dependency>
+    <dependencies>
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.16.16</version>
+            <scope>provided</scope>
+        </dependency>
 
-    <dependency>
-      <groupId>io.vavr</groupId>
-      <artifactId>vavr</artifactId>
-      <version>0.9.0</version>
-    </dependency>
+        <dependency>
+            <groupId>io.vavr</groupId>
+            <artifactId>vavr</artifactId>
+            <version>0.9.0</version>
+        </dependency>
 
-    <!-- Server -->
-	<dependency>
-	  <groupId>org.jooby</groupId>
-	  <artifactId>jooby-undertow</artifactId>
-	</dependency>
+        <!-- Server -->
+        <dependency>
+            <groupId>org.jooby</groupId>
+            <artifactId>jooby-undertow</artifactId>
+        </dependency>
 
-    <!-- logging -->
-    <dependency>
-      <groupId>ch.qos.logback</groupId>
-      <artifactId>logback-classic</artifactId>
-    </dependency>
+        <!-- logging -->
+        <dependency>
+            <groupId>ch.qos.logback</groupId>
+            <artifactId>logback-classic</artifactId>
+        </dependency>
 
-    <!-- Jackson -->
-    <dependency>
-      <groupId>org.jooby</groupId>
-      <artifactId>jooby-jackson</artifactId>
-    </dependency>
+        <!-- Jackson -->
+        <dependency>
+            <groupId>org.jooby</groupId>
+            <artifactId>jooby-jackson</artifactId>
+        </dependency>
 
-    <dependency>
-      <groupId>com.fasterxml.jackson.module</groupId>
-      <artifactId>jackson-module-parameter-names</artifactId>
-      <version>2.8.9</version>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.datatype</groupId>
-      <artifactId>jackson-datatype-jdk8</artifactId>
-      <version>2.8.9</version>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.datatype</groupId>
-      <artifactId>jackson-datatype-jsr310</artifactId>
-      <version>2.8.9</version>
-    </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.module</groupId>
+            <artifactId>jackson-module-parameter-names</artifactId>
+            <version>${jackson.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.datatype</groupId>
+            <artifactId>jackson-datatype-jdk8</artifactId>
+            <version>${jackson.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.datatype</groupId>
+            <artifactId>jackson-datatype-jsr310</artifactId>
+            <version>${jackson.version}</version>
+        </dependency>
 
-	<!-- Data -->
-    <dependency>
-      <groupId>org.jooby</groupId>
-      <artifactId>jooby-mongodb</artifactId>
-    </dependency>
 
-	<dependency>
-	  <groupId>org.jooby</groupId>
-	  <artifactId>jooby-jongo</artifactId>
-	</dependency>
-		
-	
-    <!-- Tests -->
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <scope>test</scope>
-    </dependency>
+        <!-- Data -->
+        <dependency>
+            <groupId>org.jooby</groupId>
+            <artifactId>jooby-mongodb</artifactId>
+        </dependency>
 
-    <dependency>
-      <groupId>io.rest-assured</groupId>
-      <artifactId>rest-assured</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
+        <dependency>
+            <groupId>org.jooby</groupId>
+            <artifactId>jooby-jongo</artifactId>
+        </dependency>
 
-  <build>
-    <plugins>
-      <!-- Build fat jar -->
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-shade-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
+
+        <!-- Tests -->
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>io.rest-assured</groupId>
+            <artifactId>rest-assured</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <plugins>
+            <!-- Build fat jar -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-shade-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
