+++ src/main/java/fr/avalonlab/blog/workshop/blogpost/BlogPostRepository.java
@@ -1,4 +1,4 @@
-package info.lefoll.jeff.polymer.workshop.comment;
+package fr.avalonlab.blog.workshop.blogpost;
 
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
@@ -17,15 +17,17 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 
+import static org.jongo.Oid.withOid;
+
 @Singleton
-public class CommentRepository {
+public class BlogPostRepository {
 
     private final Logger log = LoggerFactory.getLogger(getClass());
 
     private Jongo jongo;
 
     @Inject
-    public CommentRepository(@Named("dbName") String dbName, MongoClient mongoClient) {
+    public BlogPostRepository(@Named("dbName") String dbName, MongoClient mongoClient) {
         DB mongoDb = mongoClient.getDB(dbName);
 
         jongo = new Jongo(mongoDb,
@@ -36,23 +38,31 @@ public class CommentRepository {
     }
 
     private MongoCollection getCollection() {
-        return  jongo.getCollection("comments");
+        return  jongo.getCollection("blogposts");
     }
 
-    public List<Comment> findAll(String blogId) {
-        MongoCursor<Comment> result = getCollection().find("{blogId: #}", blogId).as(Comment.class);
+    public List<BlogPost> findAll() {
+        MongoCursor<BlogPost> result = getCollection().find().as(BlogPost.class);
 
-        return List.ofAll(result::iterator);
+        return List.ofAll(result);
     }
 
-    public Comment findById(String id) {
+
+    public BlogPost findById(String id) {
         ObjectId query = new ObjectId(id);
 
-        return getCollection().findOne(query).as(Comment.class);
+        return getCollection().findOne(query).as(BlogPost.class);
     }
 
-    public Comment create(Comment entity) {
+    public BlogPost create(BlogPost entity) {
         getCollection().save(entity);
+
+        return entity;
+    }
+
+    public BlogPost update(BlogPost entity) {
+
+        getCollection().update(withOid(entity.get_id())).with(entity);
 
         return entity;
     }
@@ -61,5 +71,15 @@ public class CommentRepository {
         ObjectId query = new ObjectId(id);
 
         getCollection().remove(query);
+    }
+
+    public void removeAll(){
+        getCollection().drop();
+    }
+
+    public List<BlogPost> find(String field, String param) {
+        MongoCursor<BlogPost> result = getCollection().find("{" + field +": #}", param).as(BlogPost.class);
+
+        return List.ofAll(result);
     }
 }
