+++ src/test/java/fr/avalonlab/blog/workshop/AppTest.java
@@ -1,11 +1,10 @@
-package info.lefoll.jeff.polymer.workshop;
+package fr.avalonlab.blog.workshop;
 
 import static io.restassured.RestAssured.get;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertEquals;
 
 import org.jooby.test.JoobyRule;
-import org.jooby.test.MockRouter;
 import org.junit.ClassRule;
 import org.junit.Test;
